Refacto Java JWT code - step 4 - Move dedicated user code to user repository
