+++ services/users/src/main/java/org/opfab/users/configuration/jwt/groups/GroupsMode.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018-2021, RTE (http://www.rte-france.com)
+/* Copyright (c) 2018-2025, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,14 +7,14 @@
  * This file is part of the OperatorFabric project.
  */
 
-
-package org.opfab.springtools.configuration.oauth.jwt.groups;
+package org.opfab.users.configuration.jwt.groups;
 
 /**
- * Define the 2 modes how to retrieve the user role : OPERATOR_FABRIC mode or JWT mode.
+ * Define the 2 modes how to retrieve the user role : OPERATOR_FABRIC mode or
+ * JWT mode.
  *
  *
  */
 public enum GroupsMode {
-	OPERATOR_FABRIC, JWT
+    OPERATOR_FABRIC, JWT
 }
