+++ services/users/src/test/java/org/opfab/users/configuration/jwt/groups/GroupsUtilsShould.java
@@ -7,7 +7,7 @@
  * This file is part of the OperatorFabric project.
  */
 
-package org.opfab.springtools.configuration.oauth.jwt.groups;
+package org.opfab.users.configuration.jwt.groups;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
