+++ services/users/src/test/java/org/opfab/users/configuration/jwt/groups/TestApplication.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018-2023, RTE (http://www.rte-france.com)
+/* Copyright (c) 2018-2025, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,32 +7,24 @@
  * This file is part of the OperatorFabric project.
  */
 
-package org.opfab.springtools.configuration.oauth.jwt.groups;
+package org.opfab.users.configuration.jwt.groups;
 
 import org.opfab.springtools.configuration.oauth.jwt.JwtProperties;
-import org.opfab.springtools.configuration.oauth.jwt.groups.GroupsProperties;
-import org.opfab.springtools.configuration.oauth.jwt.groups.GroupsUtils;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Import;
 
-
-
 @SpringBootApplication
 @EnableConfigurationProperties
-@Import({GroupsProperties.class, GroupsUtils.class, JwtProperties.class})
+@Import({ GroupsProperties.class, GroupsUtils.class, JwtProperties.class })
 public class TestApplication {
 
-
-    
     public static void main(String[] args) {
         ConfigurableApplicationContext ctx = SpringApplication.run(TestApplication.class, args);
 
         assert (ctx != null);
     }
-
-
 
 }
