+++ tools/spring/spring-oauth2-utilities/build.gradle
@@ -8,5 +8,4 @@ dependencies{
     implementation project(":client:users-client-data")
     implementation project(":client:businessconfig-client-data")
     implementation project(":tools:generic:utilities")
-    testImplementation project(':tools:spring:spring-test-utilities')
 }
